requirements.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.1
Flask==3.1.3
Flask-Admin==2.0.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
python-dotenv==1.2.2
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.6
WTForms==3.2.1
